requirements.txt
Django==4.0.2
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
whitenoise==5.3.0
gunicorn==20.1.0
dj-config-url==0.1.1
python-dotenv
drf-spectacular